+++ src/test/groovy/ru/vyarus/gradle/plugin/teavm/UpstreamKitTest.groovy
@@ -3,6 +3,7 @@ package ru.vyarus.gradle.plugin.teavm
 
 import org.gradle.testkit.runner.BuildResult
 import org.gradle.testkit.runner.TaskOutcome
+import spock.lang.Ignore
 
 /**
  * @author Vyacheslav Rusakov
@@ -12,7 +13,7 @@ class UpstreamKitTest extends AbstractKitTest {
 
     String GRADLE_VERSION = '8.4'
     // https://teavm.org/maven/repository/org/teavm/teavm-classlib/
-    String TEAVM_RECENT = '0.9.0-dev-15'
+    String TEAVM_RECENT = '0.10.0-dev-5'
 
     def "Check plugin execution for the latest gradle"() {
         setup:
@@ -47,7 +48,7 @@ public class Main {
         result.output.contains('Output file successfully built')
     }
 
-
+    @Ignore // todo no compatible dev versions yet
     def "Check plugin execution with recent compiler"() {
         setup:
         build """