+++ src/test/groovy/ru/vyarus/gradle/plugin/teavm/VersionDetectionKitTest.groovy
@@ -2,6 +2,7 @@ package ru.vyarus.gradle.plugin.teavm
 
 import org.gradle.testkit.runner.BuildResult
 import org.gradle.testkit.runner.TaskOutcome
+import spock.lang.Ignore
 import spock.lang.IgnoreIf
 
 /**
@@ -12,6 +13,7 @@ import spock.lang.IgnoreIf
 @IgnoreIf({ System.getenv('APPVEYOR') })
 class VersionDetectionKitTest extends AbstractKitTest {
 
+    @Ignore // todo no compatible dev versions yet
     def "Check teavm version auto detection from classpath"() {
         setup:
         build """
@@ -26,7 +28,7 @@ class VersionDetectionKitTest extends AbstractKitTest {
             }
             
             dependencies {
-                implementation "org.teavm:teavm-classlib:0.9.0-dev-15"
+                implementation "org.teavm:teavm-classlib:0.10.0-dev-5"
             }
 
             teavm {
@@ -52,7 +54,7 @@ public class Main {
 
         then: "task successful"
         result.task(':compileTeavm').outcome == TaskOutcome.SUCCESS
-        result.output.contains('TeaVM compiler version: 0.9.0-dev-15 (auto-detected)')
+        result.output.contains('TeaVM compiler version: 0.10.0-dev-5 (auto-detected)')
         result.output.contains('Output file successfully built')
     }
 
@@ -70,7 +72,7 @@ public class Main {
             }
             
             dependencies {
-                implementation "org.teavm:teavm-classlib:0.8.0-dev-2"
+                implementation "org.teavm:teavm-classlib:0.10.0-dev-5"
             }
 
             teavm {
@@ -96,7 +98,7 @@ public class Main {
 
         then: "task successful"
         result.task(':compileTeavm').outcome == TaskOutcome.SUCCESS
-        result.output.contains('TeaVM compiler version: 0.9.0')
+        result.output.contains('TeaVM compiler version: 0.9.1')
         result.output.contains('Output file successfully built')
     }
 }